refactor: extrai TableContainer
